feat: related show grid
