+++ .eslintrc.js
@@ -119,5 +119,6 @@ module.exports = {
     SignPad: true,
     RF: true,
     $select2MatcherAll: true,
+    $isFullUrl: true,
   },
 }
