+++ src/main/resources/web/admin/audit/login-logs.html
@@ -45,6 +45,7 @@
       </div>
     </div>
     <th:block th:replace="~{/_include/footer}" />
+    <script th:src="@{/assets/js/rb-datalist.common.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-datalist.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/admin/login-logs.js}" type="text/babel"></script>
   </body>
