+++ src/main/resources/web/admin/audit/recycle-bin.html
@@ -48,6 +48,7 @@
       </div>
     </div>
     <th:block th:replace="~{/_include/footer}" />
+    <script th:src="@{/assets/js/rb-datalist.common.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-datalist.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/admin/recycle-bin.js}" type="text/babel"></script>
   </body>
