+++ src/main/resources/web/admin/audit/revision-history.html
@@ -59,6 +59,7 @@
       </div>
     </div>
     <th:block th:replace="~{/_include/footer}" />
+    <script th:src="@{/assets/js/rb-datalist.common.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-datalist.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/admin/revision-history.js}" type="text/babel"></script>
   </body>
