+++ src/main/resources/web/assets/js/rb-datalist.js
@@ -4,899 +4,9 @@ Copyright (c) REBUILD <https://getrebuild.com/> and/or its owners. All rights re
 rebuild is dual-licensed under commercial and open source licenses (GPLv3).
 See LICENSE and COMMERCIAL in the project root for license information.
 */
+// @see rb-datalist.common.js
 
 const wpc = window.__PageConfig || {}
-
-const COLUMN_MIN_WIDTH = 30
-const COLUMN_MAX_WIDTH = 800
-const COLUMN_DEF_WIDTH = 130
-
-// IE/Edge 不支持首/列固定
-const supportFixedColumns = !($.browser.msie || $.browser.msedge) && $(window).width() > 767
-
-// ~~ 数据列表
-
-class RbList extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.__defaultFilterKey = `AdvFilter-${props.config.entity}`
-    this.__sortFieldKey = `SortField-${props.config.entity}`
-    this.__columnWidthKey = `ColumnWidth-${props.config.entity}.`
-
-    const sort = ($storage.get(this.__sortFieldKey) || props.config.sort || ':').split(':')
-    const fields = props.config.fields || []
-    for (let i = 0; i < fields.length; i++) {
-      const cw = $storage.get(this.__columnWidthKey + fields[i].field)
-      if (!!cw && ~~cw >= COLUMN_MIN_WIDTH) fields[i].width = ~~cw
-
-      if (sort[0] === fields[i].field) fields[i].sort = sort[1]
-      if (['SIGN', 'N2NREFERENCE', 'MULTISELECT', 'FILE', 'IMAGE', 'AVATAR'].includes(fields[i].type)) fields[i].unsort = true
-    }
-
-    delete props.config.fields
-    this.state = { ...props, fields: fields, rowsData: [], pageNo: 1, pageSize: 20, inLoad: true }
-
-    this.__defaultColumnWidth = $('#react-list').width() / 10
-    if (this.__defaultColumnWidth < COLUMN_DEF_WIDTH) this.__defaultColumnWidth = COLUMN_DEF_WIDTH
-
-    this.pageNo = 1
-    this.pageSize = $storage.get('ListPageSize') || 20
-    this.advFilterId = wpc.advFilter !== true ? null : $storage.get(this.__defaultFilterKey) // 无高级查询
-    this.fixedColumns = supportFixedColumns && this.props.uncheckbox !== true
-  }
-
-  render() {
-    const that = this
-    const lastIndex = this.state.fields.length
-
-    return (
-      <React.Fragment>
-        <div className="row rb-datatable-body">
-          <div className="col-sm-12">
-            <div className="rb-scroller" ref={(c) => (this._rblistScroller = c)}>
-              <table className="table table-hover table-striped">
-                <thead>
-                  <tr>
-                    {this.props.uncheckbox !== true && (
-                      <th className={`column-checkbox ${supportFixedColumns ? 'column-fixed' : ''}`}>
-                        <div>
-                          <label className="custom-control custom-control-sm custom-checkbox">
-                            <input className="custom-control-input" type="checkbox" onChange={(e) => this._toggleRows(e)} ref={(c) => (this._checkAll = c)} />
-                            <i className="custom-control-label" />
-                          </label>
-                        </div>
-                      </th>
-                    )}
-                    {this.state.fields.map((item, idx) => {
-                      const cWidth = item.width || that.__defaultColumnWidth
-                      const styles = { width: cWidth }
-                      const clazz = `unselect sortable ${idx === 0 && this.fixedColumns ? 'column-fixed column-fixed-2nd' : ''}`
-                      return (
-                        <th key={`column-${item.field}`} style={styles} className={clazz} data-field={item.field} onClick={(e) => !item.unsort && this._sortField(item.field, e)}>
-                          <div style={styles}>
-                            <span style={{ width: cWidth - 8 }}>{item.label}</span>
-                            <i className={`zmdi ${item.sort || ''}`} />
-                            <i className="dividing" />
-                          </div>
-                        </th>
-                      )
-                    })}
-                    <th className="column-empty" />
-                  </tr>
-                </thead>
-                <tbody ref={(c) => (this._rblistBody = c)}>
-                  {this.state.rowsData.map((item) => {
-                    const lastPrimary = item[lastIndex]
-                    const rowKey = `row-${lastPrimary.id}`
-                    return (
-                      <tr key={rowKey} data-id={lastPrimary.id} onClick={(e) => this._clickRow(e, true)}>
-                        {this.props.uncheckbox !== true && (
-                          <td key={`${rowKey}-checkbox`} className={`column-checkbox ${supportFixedColumns ? 'column-fixed' : ''}`}>
-                            <div>
-                              <label className="custom-control custom-control-sm custom-checkbox">
-                                <input className="custom-control-input" type="checkbox" onChange={(e) => this._clickRow(e)} />
-                                <i className="custom-control-label" />
-                              </label>
-                            </div>
-                          </td>
-                        )}
-                        {item.map((cell, index) => {
-                          return that.renderCell(cell, index, lastPrimary)
-                        })}
-                        <td className="column-empty" />
-                      </tr>
-                    )
-                  })}
-                </tbody>
-              </table>
-              {this.state.inLoad === false && this.state.rowsData.length === 0 && (
-                <div className="list-nodata">
-                  <span className="zmdi zmdi-info-outline" />
-                  <p>{$L('暂无数据')}</p>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-        {this.state.rowsData.length > 0 && <RbListPagination ref={(c) => (this._Pagination = c)} pageSize={this.pageSize} $$$parent={this} />}
-        {this.state.inLoad === true && <RbSpinner />}
-      </React.Fragment>
-    )
-  }
-
-  componentDidMount() {
-    // FIXME 拖到底部才能左右滚动
-
-    const $scroller = $(this._rblistScroller)
-    $scroller.perfectScrollbar({
-      // suppressScrollY: true,
-      // wheelSpeed: 1,
-    })
-
-    // use pin
-    if ($(window).height() > 666 && $(window).width() >= 1280) {
-      $('.main-content').addClass('pb-0')
-      if (supportFixedColumns) $scroller.find('.table').addClass('table-header-fixed')
-
-      $addResizeHandler(() => {
-        let mh = $(window).height() - 210
-        if ($('.main-content>.nav-tabs-classic').length > 0) mh -= 38 // Has tab
-        if ($('.main-content .quick-filter-pane').length > 0) mh -= 84 // Has query-pane
-        $scroller.css({ maxHeight: mh })
-        $scroller.perfectScrollbar('update')
-      })()
-    } else {
-      $('.main-content .rb-datatable-header').addClass('header-fixed')
-    }
-
-    if (supportFixedColumns) {
-      let slLast = 0
-      $scroller.on('ps-scroll-x', () => {
-        const sl = $scroller[0].scrollLeft
-        if (sl === slLast) return
-        slLast = sl
-        if (sl > 0) $scroller.addClass('column-fixed-pin')
-        else $scroller.removeClass('column-fixed-pin')
-      })
-    }
-
-    const that = this
-    $scroller.find('th .dividing').draggable({
-      containment: '.rb-datatable-body',
-      axis: 'x',
-      helper: 'clone',
-      start: function () {
-        that.__columnResizing = true
-      },
-      stop: function (event, ui) {
-        const field = $(event.target).parents('th').data('field')
-        let left = ui.position.left - 0
-        if (left < COLUMN_MIN_WIDTH) left = COLUMN_MIN_WIDTH
-        else if (left > COLUMN_MAX_WIDTH) left = COLUMN_MAX_WIDTH
-        const fields = that.state.fields
-        for (let i = 0; i < fields.length; i++) {
-          if (fields[i].field === field) {
-            fields[i].width = left
-            $storage.set(that.__columnWidthKey + field, left)
-            break
-          }
-        }
-        that.setState({ fields: fields }, () => $scroller.perfectScrollbar('update'))
-        setTimeout(() => (that.__columnResizing = false), 100)
-      },
-    })
-
-    // 首次由 AdvFilter 加载
-    if (wpc.advFilter !== true) this.fetchList(this._buildQuick())
-
-    $(document).on('keydown', (e) => this._keyEvent(e))
-  }
-
-  fetchList(filter) {
-    const fields = []
-    let fieldSort = null
-    this.state.fields.forEach(function (item) {
-      fields.push(item.field)
-      if (item.sort) fieldSort = `${item.field}:${item.sort.replace('sort-', '')}`
-    })
-    this.lastFilter = filter || this.lastFilter
-
-    const entity = this.props.config.entity
-    const reload = this._forceReload || this.pageNo === 1
-    this._forceReload = false
-
-    const query = {
-      entity: entity,
-      fields: fields,
-      pageNo: this.pageNo,
-      pageSize: this.pageSize,
-      filter: this.lastFilter,
-      advFilter: this.advFilterId,
-      protocolFilter: wpc.protocolFilter,
-      sort: fieldSort,
-      reload: reload,
-      statsField: wpc.statsField === true && rb.commercial > 0,
-    }
-    this.__lastQueryEntry = query
-
-    const loadingTimer = setTimeout(() => {
-      this.setState({ inLoad: true }, () => $('#react-list').addClass('rb-loading-active'))
-    }, 400)
-
-    $.post(`/app/${entity}/data-list`, JSON.stringify(query), (res) => {
-      if (res.error_code === 0) {
-        this.setState({ rowsData: res.data.data || [], inLoad: false }, () => {
-          RbList.renderAfter()
-          this._clearSelected()
-          $(this._rblistScroller).scrollTop(0)
-        })
-
-        if (reload && this._Pagination) {
-          this._Pagination.setState({ rowsTotal: res.data.total, rowsStats: res.data.stats, pageNo: this.pageNo })
-        }
-      } else {
-        RbHighbar.error(res.error_msg)
-      }
-
-      clearTimeout(loadingTimer)
-      $('#react-list').removeClass('rb-loading-active')
-    })
-  }
-
-  // 渲染表格及相关事件处理
-
-  renderCell(cellVal, index, lastPrimary) {
-    if (this.state.fields.length === index) return null
-    const field = this.state.fields[index]
-    if (!field) return null
-
-    const cellKey = `row-${lastPrimary.id}-${index}`
-    const width = this.state.fields[index].width || this.__defaultColumnWidth
-    let type = field.type
-    if (field.field === this.props.config.nameField) {
-      cellVal = lastPrimary
-      type = '$NAME$'
-    } else if (cellVal === '$NOPRIVILEGES$') {
-      type = cellVal
-    }
-
-    const c = CellRenders.render(cellVal, type, width, `${cellKey}.${field.field}`)
-    if (index === 0 && this.fixedColumns) {
-      return React.cloneElement(c, { className: `${c.props.className || ''} column-fixed column-fixed-2nd` })
-    }
-    return c
-  }
-
-  // 全选
-  _toggleRows(e, uncheck) {
-    const $body = $(this._rblistBody)
-    if (e.target.checked) {
-      $body.find('>tr').addClass('active').find('.custom-control-input').prop('checked', true)
-    } else {
-      $body.find('>tr').removeClass('active').find('.custom-control-input').prop('checked', false)
-    }
-    if (!uncheck) this._checkSelected()
-  }
-
-  // 单选
-  _clickRow(e, unhold) {
-    const $target = $(e.target)
-    if ($target.hasClass('custom-control-label')) return
-    if ($target.hasClass('custom-control-input') && unhold) return
-
-    const $tr = $target.parents('tr')
-    if (unhold) {
-      this._toggleRows({ target: { checked: false } }, true)
-      $tr.addClass('active').find('.custom-control-input').prop('checked', true)
-    } else {
-      if (e.target.checked) $tr.addClass('active')
-      else $tr.removeClass('active')
-    }
-
-    this._checkSelected()
-  }
-
-  _checkSelected() {
-    const chkSelected = $(this._rblistBody).find('>tr .custom-control-input:checked').length
-
-    // 全选/半选/全清
-    const chkAll = this.state.rowsData.length
-    if (chkSelected === 0) {
-      $(this._checkAll).prop('checked', false).parent().removeClass('indeterminate')
-    } else if (chkSelected !== chkAll) {
-      $(this._checkAll).prop('checked', false).parent().addClass('indeterminate')
-    }
-
-    if (chkSelected > 0 && chkSelected === chkAll) {
-      $(this._checkAll).prop('checked', true).parent().removeClass('indeterminate')
-    }
-
-    // 操作按钮状态
-    const $oper = $('.dataTables_oper')
-    $oper.find('.J_delete, .J_view, .J_edit, .J_assign, .J_share, .J_unshare').attr('disabled', true)
-    if (chkSelected > 0) {
-      $oper.find('.J_delete, .J_assign, .J_share, .J_unshare').attr('disabled', false)
-      if (chkSelected === 1) $oper.find('.J_view, .J_edit').attr('disabled', false)
-    }
-
-    // 分页组件
-    this._Pagination && this._Pagination.setState({ selectedTotal: chkSelected })
-  }
-
-  _clearSelected() {
-    $(this._checkAll).prop('checked', false)
-    this._toggleRows({ target: { checked: false } })
-  }
-
-  // 排序
-  _sortField(field, e) {
-    if (this.__columnResizing) return // fix: firefox
-    const fields = this.state.fields
-    for (let i = 0; i < fields.length; i++) {
-      if (fields[i].field === field) {
-        if (fields[i].sort === 'sort-asc') fields[i].sort = 'sort-desc'
-        else if (fields[i].sort === 'sort-desc') fields[i].sort = null
-        else fields[i].sort = 'sort-asc'
-
-        if (fields[i].sort) $storage.set(this.__sortFieldKey, `${field}:${fields[i].sort}`)
-        else $storage.remove(this.__sortFieldKey)
-      } else {
-        fields[i].sort = null
-      }
-    }
-    this.setState({ fields: fields }, () => this.fetchList())
-
-    $stopEvent(e)
-    return false
-  }
-
-  _tryActive($el) {
-    if ($el.length === 1) {
-      this._clickRow({ target: $el.find('.custom-checkbox') }, true)
-    }
-  }
-
-  _keyEvent(e) {
-    if (!$(e.target).is('body')) return
-    if (!(e.keyCode === 40 || e.keyCode === 38 || e.keyCode === 13)) return
-
-    const $chk = $(this._rblistBody).find('>tr .custom-control-input:checked').last()
-    if ($chk.length === 0) return
-
-    const $tr = $chk.eq(0).parents('tr')
-    if (e.keyCode === 40) {
-      this._tryActive($tr.next())
-    } else if (e.keyCode === 38) {
-      this._tryActive($tr.prev())
-    } else {
-      $('.J_view').trigger('click')
-    }
-  }
-
-  // 外部接口
-
-  /**
-   * 分页设置
-   */
-  setPage(pageNo, pageSize) {
-    this.pageNo = pageNo || this.pageNo
-    if (pageSize) {
-      this.pageSize = pageSize
-      $storage.set('ListPageSize', pageSize)
-    }
-    this.fetchList()
-  }
-
-  /**
-   * 设置高级过滤器 ID
-   */
-  setAdvFilter(id) {
-    this.advFilterId = id
-    this.pageNo = 1
-    this.fetchList(this._buildQuick())
-    if (id) $storage.set(this.__defaultFilterKey, id)
-    else $storage.remove(this.__defaultFilterKey)
-  }
-
-  /**
-   * 重新加载
-   */
-  reload() {
-    this._forceReload = true
-    this.fetchList()
-  }
-
-  /**
-   * 搜索
-   */
-  search(filter, fromAdv) {
-    this.pageNo = 1
-    this.fetchList(filter)
-
-    // 高级查询
-    if (fromAdv === true) {
-      $('.J_advfilter .indicator-primary').remove()
-      if (filter.items.length > 0) $('<i class="indicator-primary bg-warning"></i>').appendTo('.J_advfilter')
-    }
-  }
-
-  // @el - search element
-  searchQuick = (el) => this.search(this._buildQuick(el))
-
-  _buildQuick(el) {
-    el = $(el || '.input-search>input')
-    const q = el.val()
-    if (!q && !this.lastFilter) return null
-
-    return {
-      entity: this.props.config.entity,
-      type: 'QUICK',
-      values: { 1: q },
-      quickFields: el.data('quickfields') || null,
-    }
-  }
-
-  /**
-   * 获取选中 ID[]
-   */
-  getSelectedIds(noWarn) {
-    const selected = []
-    $(this._rblistBody)
-      .find('>tr .custom-control-input:checked')
-      .each(function () {
-        selected.push($(this).parents('tr').data('id'))
-      })
-
-    if (selected.length === 0 && noWarn !== true) RbHighbar.create($L('未选中任何记录'))
-    return selected
-  }
-
-  /**
-   * 获取最后查询记录总数
-   */
-  getLastQueryTotal() {
-    return this._Pagination ? this._Pagination.state.rowsTotal : 0
-  }
-
-  /**
-   * 获取最后查询条件
-   */
-  getLastQueryEntry() {
-    return JSON.parse(JSON.stringify(this.__lastQueryEntry)) // Use clone
-  }
-
-  // 渲染完成后回调
-  static renderAfter() {}
-}
-
-function _isFullUrl(urlKey) {
-  return urlKey.startsWith('http://') || urlKey.startsWith('https://')
-}
-
-// 列表（单元格）渲染
-const CellRenders = {
-  // 打开记录
-  clickView(v, e) {
-    RbViewModal.create({ id: v.id, entity: v.entity })
-    e && $stopEvent(e)
-    return false
-  },
-
-  // 打开预览
-  clickPreview(v, idx, e) {
-    RbPreview.create(v, idx)
-    e && $stopEvent(e)
-    return false
-  },
-
-  __renders: {},
-
-  addRender(type, func) {
-    this.__renders[type] = func
-  },
-
-  render(value, type, width, key) {
-    const style = { width: width || COLUMN_MIN_WIDTH }
-    if (!value) return this.renderSimple(value, style, key)
-    else return (this.__renders[type] || this.renderSimple)(value, style, key)
-  },
-
-  /**
-   * @param {*} v 值
-   * @param {*} s 样式
-   * @param {*} k key of React (contains fieldName)
-   */
-  renderSimple(v, s, k) {
-    if (typeof v === 'string' && v.length > 300) v = v.substr(0, 300)
-    else if (k.endsWith('.approvalId') && !v) v = $L('未提交')
-    else if (k.endsWith('.approvalState') && !v) v = $L('草稿')
-
-    return (
-      <td key={k}>
-        <div style={s} title={typeof v === 'string' ? v : null}>
-          {v || ''}
-        </div>
-      </td>
-    )
-  },
-}
-
-// 名称字段
-CellRenders.addRender('$NAME$', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v.text}>
-        <a href={`#!/View/${v.entity}/${v.id}`} onClick={(e) => CellRenders.clickView(v, e)} className="column-main">
-          {v.text}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-// 无权访问字段
-CellRenders.addRender('$NOPRIVILEGES$', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} className="column-nopriv">
-        [{$L('无权限')}]
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('IMAGE', function (v, s, k) {
-  v = v || []
-  const vLen = v.length
-  return (
-    <td key={k} className="td-sm" title={$L('共 %d 项', vLen)}>
-      <div className="column-imgs" style={s}>
-        {v.map((item, idx) => {
-          const imgName = $fileCutName(item)
-          const imgUrl = _isFullUrl(item) ? item : `${rb.baseUrl}/filex/img/${item}`
-          return (
-            <a key={item} title={imgName} onClick={(e) => CellRenders.clickPreview(v, idx, e)}>
-              <img alt="IMG" src={`${imgUrl}?imageView2/2/w/100/interlace/1/q/100`} />
-            </a>
-          )
-        })}
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('FILE', function (v, s, k) {
-  v = v || []
-  const vLen = v.length
-  return (
-    <td key={k} title={$L('共 %d 项', vLen)}>
-      <div className="column-files" style={s}>
-        {v.map((item) => {
-          const fileName = $fileCutName(item)
-          return (
-            <a key={item} title={fileName} onClick={(e) => CellRenders.clickPreview(item, null, e)}>
-              {fileName}
-            </a>
-          )
-        })}
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('REFERENCE', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v.text}>
-        <a href={`#!/View/${v.entity}/${v.id}`} onClick={(e) => CellRenders.clickView(v, e)}>
-          {v.text}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('N2NREFERENCE', function (v, s, k) {
-  v = v || []
-  const vLen = v.length
-  return (
-    <td key={k} className="td-sm" title={$L('共 %d 项', vLen)}>
-      <div className="column-multi" style={s}>
-        {v.map((item) => {
-          return (
-            <a key={item.id} title={item.text} className="badge hover-color" href={`#!/View/${item.entity}/${item.id}`} onClick={(e) => CellRenders.clickView(item, e)}>
-              {item.text}
-            </a>
-          )
-        })}
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('URL', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v}>
-        <a href={`${rb.baseUrl}/commons/url-safe?url=${$encode(v)}`} className="column-url" target="_blank" rel="noopener noreferrer" onClick={(e) => $stopEvent(e)}>
-          {v}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('EMAIL', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v}>
-        <a href={`mailto:${v}`} className="column-url" onClick={(e) => $stopEvent(e)}>
-          {v}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('PHONE', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v}>
-        <a href={`tel:${v}`} className="column-url" onClick={(e) => $stopEvent(e)}>
-          {v}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-const APPROVAL_STATE_CLAZZs = {
-  [$L('审批中')]: 'warning',
-  [$L('驳回')]: 'danger',
-  [$L('通过')]: 'success',
-}
-CellRenders.addRender('STATE', function (v, s, k) {
-  if (k.endsWith('.approvalState')) {
-    const badge = APPROVAL_STATE_CLAZZs[v]
-    return (
-      <td key={k} className="td-sm column-state">
-        <div style={s} title={v}>
-          <span className={badge ? `badge badge-${badge}` : ''}>{v}</span>
-        </div>
-      </td>
-    )
-  } else {
-    return CellRenders.renderSimple(v, s, k)
-  }
-})
-
-const renderNumber = function (v, s, k) {
-  if ((v + '').substr(0, 1) === '-') {
-    return (
-      <td key={k}>
-        <div className="text-danger" style={s} title={v}>
-          {v}
-        </div>
-      </td>
-    )
-  } else {
-    return CellRenders.renderSimple(v, s, k)
-  }
-}
-CellRenders.addRender('DECIMAL', renderNumber)
-CellRenders.addRender('NUMBER', renderNumber)
-
-CellRenders.addRender('MULTISELECT', function (v, s, k) {
-  const vLen = (v.text || []).length
-  return (
-    <td key={k} className="td-sm" title={$L('共 %d 项', vLen)}>
-      <div className="column-multi" style={s}>
-        {(v.text || []).map((item) => {
-          if (typeof item === 'object') {
-            const style2 = { borderColor: item.color, backgroundColor: item.color, color: '#fff' }
-            // const style2 = { borderColor: item.color, color: item.color }
-            return (
-              <span key={item.text} className="badge" title={item.text} style={style2}>
-                {item.text}
-              </span>
-            )
-          } else {
-            return (
-              <span key={item} className="badge" title={item}>
-                {item}
-              </span>
-            )
-          }
-        })}
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('AVATAR', function (v, s, k) {
-  const imgUrl = _isFullUrl(v) ? v : `${rb.baseUrl}/filex/img/${v}?imageView2/2/w/100/interlace/1/q/100`
-  return (
-    <td key={k} className="user-avatar">
-      <img src={imgUrl} alt="Avatar" />
-    </td>
-  )
-})
-
-CellRenders.addRender('LOCATION', function (v, s, k) {
-  return (
-    <td key={k}>
-      <div style={s} title={v.text}>
-        <a
-          href={`#!/Map:${v.lng || ''},${v.lat || ''}`}
-          onClick={(e) => {
-            $stopEvent(e, true)
-            BaiduMapModal.view(v)
-          }}>
-          {v.text}
-        </a>
-      </div>
-    </td>
-  )
-})
-
-CellRenders.addRender('SIGN', function (v, s, k) {
-  return (
-    <td key={k} className="user-avatar sign">
-      <img alt="SIGN" src={v} />
-    </td>
-  )
-})
-
-CellRenders.addRender('PICKLIST', function (v, s, k) {
-  // Use badge
-  if (typeof v === 'object') {
-    const style2 = { borderColor: v.color, backgroundColor: v.color, color: '#fff' }
-    // const style2 = { borderColor: v.color, color: v.color, backgroundColor: '#fff' }
-    return (
-      <td key={k} className="td-sm column-state">
-        <div style={s} title={v.text}>
-          <span className="badge" style={style2}>
-            {v.text}
-          </span>
-        </div>
-      </td>
-    )
-  } else {
-    return CellRenders.renderSimple(v, s, k)
-  }
-})
-
-// ~ 分页组件
-
-class RbListPagination extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = { ...props, pageNo: props.pageNo || 1, pageSize: props.pageSize || 20, rowsTotal: props.rowsTotal || 0 }
-    // via List
-    this._entity = this.props.$$$parent.props.config.entity
-  }
-
-  render() {
-    this.__pageTotal = Math.ceil(this.state.rowsTotal / this.state.pageSize)
-    if (this.__pageTotal <= 0) this.__pageTotal = 1
-    const pages = this.__pageTotal <= 1 ? [1] : $pages(this.__pageTotal, this.state.pageNo)
-
-    return (
-      <div className="row rb-datatable-footer">
-        <div className="col-12 col-lg-6 col-xl-7">
-          <div className="dataTables_info" key="page-rowsTotal">
-            {this._renderStats()}
-          </div>
-        </div>
-        <div className="col-12 col-lg-6 col-xl-5">
-          <div className="float-right paging_sizes">
-            <select className="form-control form-control-sm" title={$L('每页显示')} onChange={this.setPageSize} value={this.state.pageSize || 20}>
-              {rb.env === 'dev' && <option value="5">5</option>}
-              <option value="20">20</option>
-              <option value="40">40</option>
-              <option value="80">80</option>
-              <option value="100">100</option>
-              <option value="200">200</option>
-              <option value="300">300</option>
-              <option value="400">400</option>
-              <option value="500">500</option>
-            </select>
-          </div>
-          <div className="float-right dataTables_paginate paging_simple_numbers">
-            <ul className="pagination mb-0">
-              {this.state.pageNo > 1 && (
-                <li className="paginate_button page-item">
-                  <a className="page-link" onClick={() => this.prev()}>
-                    <span className="icon zmdi zmdi-chevron-left" />
-                  </a>
-                </li>
-              )}
-              {pages.map((item, idx) => {
-                if (item === '.')
-                  return (
-                    <li key={`pnx-${idx}`} className="paginate_button page-item disabled">
-                      <a className="page-link">...</a>
-                    </li>
-                  )
-                else
-                  return (
-                    <li key={`pn-${item}`} className={`paginate_button page-item ${this.state.pageNo === item && 'active'}`}>
-                      <a className="page-link" onClick={this.goto.bind(this, item)}>
-                        {item}
-                      </a>
-                    </li>
-                  )
-              })}
-              {this.state.pageNo !== this.__pageTotal && (
-                <li className="paginate_button page-item">
-                  <a className="page-link" onClick={() => this.next()}>
-                    <span className="icon zmdi zmdi-chevron-right" />
-                  </a>
-                </li>
-              )}
-            </ul>
-          </div>
-          <div className="clearfix" />
-        </div>
-      </div>
-    )
-  }
-
-  _renderStats() {
-    return (
-      <div>
-        {this.state.selectedTotal > 0 && <span className="mr-1">{$L('已选中 %d 条', this.state.selectedTotal)}.</span>}
-        {this.state.rowsTotal > 0 && <span>{$L('共 %d 条数据', this.state.rowsTotal)}</span>}
-        {(this.state.rowsStats || []).map((item, idx) => {
-          return (
-            <span key={idx} className="stat-item">
-              {item.label} <strong className="text-warning">{item.value} </strong>
-            </span>
-          )
-        })}
-        {rb.isAdminUser && wpc.statsField && (
-          <a
-            className="list-stats-settings"
-            onClick={() =>
-              RbModal.create(
-                `/p/admin/metadata/list-stats?entity=${this._entity}`,
-                <React.Fragment>
-                  {$L('配置统计字段')}
-                  <sup className="rbv" title={$L('增值功能')} />
-                </React.Fragment>
-              )
-            }>
-            <i className="icon zmdi zmdi-settings" title={$L('配置统计字段')} />
-          </a>
-        )}
-      </div>
-    )
-  }
-
-  prev() {
-    if (this.state.pageNo === 1) return
-    this.goto(this.state.pageNo - 1)
-  }
-
-  next() {
-    if (this.state.pageNo === this.__pageTotal) return
-    this.goto(this.state.pageNo + 1)
-  }
-
-  goto(pageNo) {
-    this.setState({ pageNo: pageNo }, () => {
-      this.props.$$$parent.setPage(this.state.pageNo)
-    })
-  }
-
-  setPageSize = (e) => {
-    const s = e.target.value
-    this.setState({ pageSize: s, pageNo: 1 }, () => {
-      this.props.$$$parent.setPage(1, s)
-    })
-  }
-}
 
 // ~~ 列表操作
