+++ src/main/resources/web/general/record-view.html
@@ -118,6 +118,7 @@
         viewAdds: [(${ViewAdds})],
         transformTos: [(${TransformTos})],
         recordId: '[[${id}]]',
+        forceSubView: true
       }
     </script>
     <script th:src="@{/assets/js/feeds/feeds-post.js}" type="text/babel"></script>
@@ -125,6 +126,7 @@
     <script th:src="@{/assets/js/rb-forms.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-forms.append.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-forms.protable.js}" type="text/babel"></script>
+    <script th:src="@{/assets/js/rb-datalist.common.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-view.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-view.append.js}" type="text/babel"></script>
     <script th:src="@{/assets/js/rb-assignshare.js}" type="text/babel"></script>
